feat(utils) extract and export createApiUtils - extract createApiUtils and export it to allow users to create their own api call context without the need to create a full api. This would be good for RPCs where an action is triggered upon call and no resource is strictly attached to it (such as sending an email)
