+++ lib/cjs/ts/v2/api.js
@@ -9,11 +9,12 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.createApi = exports.getPaginatedSnakeCasedZod = exports.createCustomServiceCall = void 0;
+exports.createApi = exports.createCustomServiceCall = void 0;
 const tn_utils_1 = require("@thinknimble/tn-utils");
 const zod_1 = require("zod");
-const pagination_1 = require("./pagination");
+const response_1 = require("./response");
 const utils_1 = require("./utils");
+const utils_2 = require("./utils");
 const paginationFiltersZod = zod_1.z
     .object({
     page: zod_1.z.number(),
@@ -56,47 +57,17 @@ function createCustomServiceCall(...args) {
     };
 }
 exports.createCustomServiceCall = createCustomServiceCall;
-const getSnakeCasedZodRawShape = (zodShape) => {
-    // objectToSnakeCase would mess up zod internal fields, so we must do the case conversion only on the keys of the shape
-    const unknownSnakeCasedZod = Object.fromEntries(Object.entries(zodShape).map(([k, v]) => {
-        return [(0, tn_utils_1.toSnakeCase)(k), v];
-    }));
-    return unknownSnakeCasedZod;
-};
-const getPaginatedSnakeCasedZod = (zodShape) => {
-    return (0, pagination_1.getPaginatedZod)(getSnakeCasedZodRawShape(zodShape));
-};
-exports.getPaginatedSnakeCasedZod = getPaginatedSnakeCasedZod;
 function createApi({ models, client, endpoint, }, customServiceCalls = undefined) {
     const createCustomServiceCallHandler = (serviceCallOpts, 
     /**
      * This name allow us to keep record of which method it is, so that we can identify in case of output mismatch
      */
     name) => (input) => __awaiter(this, void 0, void 0, function* () {
-        // we have to identify if we have a shape or a plain zod (only primitives are allowed). Primitive zods do not get util functions since they don't require response case transformation
-        const isInputZodPrimitive = serviceCallOpts.inputShape instanceof zod_1.z.ZodSchema;
-        const isOutputZodPrimitive = serviceCallOpts.outputShape instanceof zod_1.z.ZodSchema;
-        // since this checks for the api response, which we don't control, we can't strict parse, else we would break the flow. We'd rather safe parse and show a warning if there's a mismatch
-        const fromApi = isOutputZodPrimitive
-            ? undefined
-            : (obj) => {
-                var _a;
-                return (0, utils_1.parseResponse)({
-                    identifier: name,
-                    data: (_a = (0, tn_utils_1.objectToCamelCase)(obj)) !== null && _a !== void 0 ? _a : {},
-                    zod: zod_1.z.object(serviceCallOpts.outputShape),
-                });
-            };
-        // Given that this is under our control, we should not do safe parse, if the parsing fails means something is wrong (you're not complying with the schema you defined)
-        const toApi = isInputZodPrimitive
-            ? undefined
-            : (obj) => zod_1.z.object(getSnakeCasedZodRawShape(serviceCallOpts.inputShape)).parse((0, tn_utils_1.objectToSnakeCase)(obj));
-        // avoid leaking the wrong fields to our callback (we already protect this through ts but it does not hurt to prevent it at runtime as well)
-        const utilsResult = fromApi || toApi
-            ? {
-                utils: Object.assign(Object.assign({}, (fromApi ? { fromApi } : {})), (toApi ? { toApi } : {})),
-            }
-            : {};
+        const utilsResult = (0, utils_1.createApiUtils)({
+            name,
+            inputShape: serviceCallOpts.inputShape,
+            outputShape: serviceCallOpts.outputShape,
+        });
         const inputResult = input ? { input } : {};
         return serviceCallOpts.callback(Object.assign(Object.assign({ client,
             endpoint }, inputResult), utilsResult));
@@ -111,18 +82,18 @@ function createApi({ models, client, endpoint, }, customServiceCalls = undefined
         }
         const uri = `${endpoint}/${id}`;
         const res = yield client.get(uri);
-        const parsed = (0, utils_1.parseResponse)({
+        const parsed = (0, response_1.parseResponse)({
             identifier: `${retrieve.name} ${uri}`,
             data: res.data,
-            zod: zod_1.z.object(getSnakeCasedZodRawShape(models.entity)),
+            zod: zod_1.z.object((0, utils_2.getSnakeCasedZodRawShape)(models.entity)),
         });
         return (0, tn_utils_1.objectToCamelCase)(parsed);
     });
     const create = (inputs) => __awaiter(this, void 0, void 0, function* () {
         const snaked = (0, tn_utils_1.objectToSnakeCase)(inputs);
         const res = yield client.post(endpoint, snaked);
-        const snakedEntityShape = getSnakeCasedZodRawShape(models.entity);
-        const parsed = (0, utils_1.parseResponse)({
+        const snakedEntityShape = (0, utils_2.getSnakeCasedZodRawShape)(models.entity);
+        const parsed = (0, response_1.parseResponse)({
             identifier: `${create.name} ${endpoint}`,
             data: res.data,
             zod: zod_1.z.object(snakedEntityShape),
@@ -147,7 +118,7 @@ function createApi({ models, client, endpoint, }, customServiceCalls = undefined
                 return [[k, v]];
             }))
             : undefined;
-        const paginatedZod = (0, exports.getPaginatedSnakeCasedZod)(models.entity);
+        const paginatedZod = (0, utils_2.getPaginatedSnakeCasedZod)(models.entity);
         const res = yield client.get(endpoint, {
             params: snakedCleanParsedFilters,
         });
