+++ lib/cjs/ts/v2/index.js
@@ -1,10 +1,12 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.parseResponse = exports.createCollectionManager = exports.createCustomServiceCall = exports.createApi = void 0;
+exports.parseResponse = exports.createApiUtils = exports.createCollectionManager = exports.createCustomServiceCall = exports.createApi = void 0;
 var api_1 = require("./api");
 Object.defineProperty(exports, "createApi", { enumerable: true, get: function () { return api_1.createApi; } });
 Object.defineProperty(exports, "createCustomServiceCall", { enumerable: true, get: function () { return api_1.createCustomServiceCall; } });
 var collection_manager_1 = require("./collection-manager");
 Object.defineProperty(exports, "createCollectionManager", { enumerable: true, get: function () { return collection_manager_1.createCollectionManager; } });
 var utils_1 = require("./utils");
-Object.defineProperty(exports, "parseResponse", { enumerable: true, get: function () { return utils_1.parseResponse; } });
+Object.defineProperty(exports, "createApiUtils", { enumerable: true, get: function () { return utils_1.createApiUtils; } });
+var response_1 = require("./response");
+Object.defineProperty(exports, "parseResponse", { enumerable: true, get: function () { return response_1.parseResponse; } });
