+++ lib/cjs/types/ts/v2/index.d.ts
@@ -1,5 +1,7 @@
+export type { PaginationFilters } from "./api";
 export { createApi, createCustomServiceCall } from "./api";
 export { createCollectionManager } from "./collection-manager";
-export { parseResponse } from "./utils";
-export type { GetZodInferredTypeFromRaw } from "./api";
+export type { GetZodInferredTypeFromRaw, FromApiCall, ToApiCall } from "./utils";
+export { createApiUtils } from "./utils";
+export { parseResponse } from "./response";
 //# sourceMappingURL=index.d.ts.map
