+++ lib/esm/ts/v2/api.js
@@ -1,7 +1,8 @@
-import { objectToCamelCase, objectToSnakeCase, toSnakeCase, } from "@thinknimble/tn-utils";
+import { objectToCamelCase, objectToSnakeCase } from "@thinknimble/tn-utils";
 import { z } from "zod";
-import { getPaginatedZod } from "./pagination";
-import { parseResponse } from "./utils";
+import { parseResponse } from "./response";
+import { createApiUtils } from "./utils";
+import { getPaginatedSnakeCasedZod, getSnakeCasedZodRawShape } from "./utils";
 const paginationFiltersZod = z
     .object({
     page: z.number(),
@@ -43,46 +44,17 @@ export function createCustomServiceCall(...args) {
         callback: second,
     };
 }
-const getSnakeCasedZodRawShape = (zodShape) => {
-    // objectToSnakeCase would mess up zod internal fields, so we must do the case conversion only on the keys of the shape
-    const unknownSnakeCasedZod = Object.fromEntries(Object.entries(zodShape).map(([k, v]) => {
-        return [toSnakeCase(k), v];
-    }));
-    return unknownSnakeCasedZod;
-};
-export const getPaginatedSnakeCasedZod = (zodShape) => {
-    return getPaginatedZod(getSnakeCasedZodRawShape(zodShape));
-};
 export function createApi({ models, client, endpoint, }, customServiceCalls = undefined) {
     const createCustomServiceCallHandler = (serviceCallOpts, 
     /**
      * This name allow us to keep record of which method it is, so that we can identify in case of output mismatch
      */
     name) => async (input) => {
-        // we have to identify if we have a shape or a plain zod (only primitives are allowed). Primitive zods do not get util functions since they don't require response case transformation
-        const isInputZodPrimitive = serviceCallOpts.inputShape instanceof z.ZodSchema;
-        const isOutputZodPrimitive = serviceCallOpts.outputShape instanceof z.ZodSchema;
-        // since this checks for the api response, which we don't control, we can't strict parse, else we would break the flow. We'd rather safe parse and show a warning if there's a mismatch
-        const fromApi = isOutputZodPrimitive
-            ? undefined
-            : (obj) => parseResponse({
-                identifier: name,
-                data: objectToCamelCase(obj) ?? {},
-                zod: z.object(serviceCallOpts.outputShape),
-            });
-        // Given that this is under our control, we should not do safe parse, if the parsing fails means something is wrong (you're not complying with the schema you defined)
-        const toApi = isInputZodPrimitive
-            ? undefined
-            : (obj) => z.object(getSnakeCasedZodRawShape(serviceCallOpts.inputShape)).parse(objectToSnakeCase(obj));
-        // avoid leaking the wrong fields to our callback (we already protect this through ts but it does not hurt to prevent it at runtime as well)
-        const utilsResult = fromApi || toApi
-            ? {
-                utils: {
-                    ...(fromApi ? { fromApi } : {}),
-                    ...(toApi ? { toApi } : {}),
-                },
-            }
-            : {};
+        const utilsResult = createApiUtils({
+            name,
+            inputShape: serviceCallOpts.inputShape,
+            outputShape: serviceCallOpts.outputShape,
+        });
         const inputResult = input ? { input } : {};
         return serviceCallOpts.callback({
             client,
