+++ lib/esm/ts/v2/index.js
@@ -1,3 +1,4 @@
 export { createApi, createCustomServiceCall } from "./api";
 export { createCollectionManager } from "./collection-manager";
-export { parseResponse } from "./utils";
+export { createApiUtils } from "./utils";
+export { parseResponse } from "./response";
