+++ lib/esm/types/ts/v2/api.d.ts
@@ -1,8 +1,8 @@
-import { SnakeCase, SnakeCasedPropertiesDeep } from "@thinknimble/tn-utils";
 import { AxiosInstance } from "axios";
-import { z, ZodRawShape } from "zod";
+import { z } from "zod";
 import { IPagination } from "../pagination";
 import { getPaginatedZod } from "./pagination";
+import { FromApiCall, GetZodInferredTypeFromRaw, ToApiCall } from "./utils";
 declare const paginationFiltersZod: z.ZodOptional<z.ZodObject<{
     page: z.ZodOptional<z.ZodNumber>;
     pageSize: z.ZodOptional<z.ZodNumber>;
@@ -21,13 +21,11 @@ declare const filtersZod: z.ZodOptional<z.ZodObject<{
 }, {
     ordering?: string | undefined;
 }>>;
-declare type ToApiCall<TInput extends ZodRawShape | z.ZodTypeAny> = (obj: object) => TInput extends z.ZodRawShape ? SnakeCasedPropertiesDeep<GetZodInferredTypeFromRaw<TInput>> : TInput extends z.ZodTypeAny ? z.infer<TInput> : never;
-declare type FromApiCall<TOutput extends ZodRawShape | z.ZodTypeAny> = (obj: object) => TOutput extends z.ZodRawShape ? GetZodInferredTypeFromRaw<TOutput> : TOutput extends z.ZodTypeAny ? z.infer<TOutput> : never;
-declare type InferCallbackInput<TInput extends ZodRawShape | z.ZodTypeAny> = TInput extends z.ZodRawShape ? GetZodInferredTypeFromRaw<TInput> : TInput extends z.ZodTypeAny ? z.infer<TInput> : never;
-declare type CallbackInput<TInput extends ZodRawShape | ZodPrimitives> = TInput extends z.ZodVoid ? unknown : {
+declare type InferCallbackInput<TInput extends z.ZodRawShape | z.ZodTypeAny> = TInput extends z.ZodRawShape ? GetZodInferredTypeFromRaw<TInput> : TInput extends z.ZodTypeAny ? z.infer<TInput> : never;
+declare type CallbackInput<TInput extends z.ZodRawShape | ZodPrimitives> = TInput extends z.ZodVoid ? unknown : {
     input: InferCallbackInput<TInput>;
 };
-declare type CallbackUtils<TInput extends ZodRawShape | ZodPrimitives, TOutput extends ZodRawShape | ZodPrimitives, TInputIsPrimitive extends boolean = TInput extends ZodPrimitives ? true : false, TOutputIsPrimitive extends boolean = TOutput extends ZodPrimitives ? true : false> = TInput extends z.ZodVoid ? TOutput extends z.ZodVoid ? unknown : TOutputIsPrimitive extends true ? unknown : {
+declare type CallbackUtils<TInput extends z.ZodRawShape | ZodPrimitives, TOutput extends z.ZodRawShape | ZodPrimitives, TInputIsPrimitive extends boolean = TInput extends ZodPrimitives ? true : false, TOutputIsPrimitive extends boolean = TOutput extends ZodPrimitives ? true : false> = TInput extends z.ZodVoid ? TOutput extends z.ZodVoid ? unknown : TOutputIsPrimitive extends true ? unknown : {
     utils: {
         fromApi: FromApiCall<TOutput>;
     };
@@ -96,29 +94,6 @@ declare type CustomServiceCallPlaceholder = {
 declare type CustomServiceCallsRecord<TOpts extends object> = TOpts extends Record<string, CustomServiceCallPlaceholder> ? {
     [TKey in keyof TOpts]: (inputs: TOpts[TKey]["inputShape"] extends z.ZodRawShape ? GetZodInferredTypeFromRaw<TOpts[TKey]["inputShape"]> : TOpts[TKey]["inputShape"] extends z.ZodTypeAny ? z.infer<TOpts[TKey]["inputShape"]> : never) => Promise<TOpts[TKey]["outputShape"] extends z.ZodRawShape ? GetZodInferredTypeFromRaw<TOpts[TKey]["outputShape"]> : TOpts[TKey]["outputShape"] extends z.ZodTypeAny ? z.infer<TOpts[TKey]["outputShape"]> : never>;
 } : never;
-declare type ZodRawShapeSnakeCased<T extends z.ZodRawShape> = {
-    [TKey in keyof T as SnakeCase<TKey>]: T[TKey];
-};
-export declare const getPaginatedSnakeCasedZod: <T extends z.ZodRawShape>(zodShape: T) => z.ZodObject<{
-    count: z.ZodNumber;
-    next: z.ZodNullable<z.ZodString>;
-    previous: z.ZodNullable<z.ZodString>;
-    results: z.ZodArray<z.ZodObject<ZodRawShapeSnakeCased<T>, "strip", z.ZodTypeAny, z.objectUtil.addQuestionMarks<ZodRawShapeSnakeCased<T> extends infer T_3 extends z.ZodRawShape ? { [k_2 in keyof T_3]: ZodRawShapeSnakeCased<T>[k_2]["_output"]; } : never> extends infer T_1 ? { [k_1 in keyof T_1]: z.objectUtil.addQuestionMarks<ZodRawShapeSnakeCased<T> extends infer T_2 extends z.ZodRawShape ? { [k in keyof T_2]: ZodRawShapeSnakeCased<T>[k]["_output"]; } : never>[k_1]; } : never, z.objectUtil.addQuestionMarks<ZodRawShapeSnakeCased<T> extends infer T_6 extends z.ZodRawShape ? { [k_2_1 in keyof T_6]: ZodRawShapeSnakeCased<T>[k_2_1]["_input"]; } : never> extends infer T_4 ? { [k_3 in keyof T_4]: z.objectUtil.addQuestionMarks<ZodRawShapeSnakeCased<T> extends infer T_5 extends z.ZodRawShape ? { [k_2 in keyof T_5]: ZodRawShapeSnakeCased<T>[k_2]["_input"]; } : never>[k_3]; } : never>, "many">;
-}, "strip", z.ZodTypeAny, {
-    next: string | null;
-    previous: string | null;
-    count: number;
-    results: (z.objectUtil.addQuestionMarks<ZodRawShapeSnakeCased<T> extends infer T_3 extends z.ZodRawShape ? { [k_2 in keyof T_3]: ZodRawShapeSnakeCased<T>[k_2]["_output"]; } : never> extends infer T_1 ? { [k_1 in keyof T_1]: z.objectUtil.addQuestionMarks<ZodRawShapeSnakeCased<T> extends infer T_2 extends z.ZodRawShape ? { [k in keyof T_2]: ZodRawShapeSnakeCased<T>[k]["_output"]; } : never>[k_1]; } : never)[];
-}, {
-    next: string | null;
-    previous: string | null;
-    count: number;
-    results: (z.objectUtil.addQuestionMarks<ZodRawShapeSnakeCased<T> extends infer T_6 extends z.ZodRawShape ? { [k_2_1 in keyof T_6]: ZodRawShapeSnakeCased<T>[k_2_1]["_input"]; } : never> extends infer T_4 ? { [k_3 in keyof T_4]: z.objectUtil.addQuestionMarks<ZodRawShapeSnakeCased<T> extends infer T_5 extends z.ZodRawShape ? { [k_2 in keyof T_5]: ZodRawShapeSnakeCased<T>[k_2]["_input"]; } : never>[k_3]; } : never)[];
-}>;
-/**
- * Get the resulting inferred type from a zod shape
- */
-export declare type GetZodInferredTypeFromRaw<T extends z.ZodRawShape> = z.infer<ReturnType<typeof z.object<T>>>;
 declare type BareApiService<TEntity extends z.ZodRawShape, TCreate extends z.ZodRawShape, TExtraFilters extends z.ZodRawShape = never> = {
     client: AxiosInstance;
     retrieve(id: string): Promise<GetZodInferredTypeFromRaw<TEntity>>;
