+++ src/ts/v2/api.test.ts
@@ -5,7 +5,8 @@ import axios from "axios"
 import { beforeEach, describe, expect, it, Mocked, vi } from "vitest"
 import { z } from "zod"
 import Pagination from "../pagination"
-import { createApi, createCustomServiceCall, getPaginatedSnakeCasedZod, GetZodInferredTypeFromRaw } from "./api"
+import { createApi, createCustomServiceCall } from "./api"
+import { getPaginatedSnakeCasedZod, GetZodInferredTypeFromRaw } from "./utils"
 
 vi.mock("axios")
