+++ src/ts/v2/index.ts
@@ -1,4 +1,6 @@
+export type { PaginationFilters } from "./api"
 export { createApi, createCustomServiceCall } from "./api"
 export { createCollectionManager } from "./collection-manager"
-export { parseResponse } from "./utils"
-export type { GetZodInferredTypeFromRaw } from "./api"
+export type { GetZodInferredTypeFromRaw, FromApiCall, ToApiCall } from "./utils"
+export { createApiUtils } from "./utils"
+export { parseResponse } from "./response"
